+++ src/pypackproxy/__init__.py
@@ -11,3 +11,5 @@ PYPP_DEBUG = bool(os.getenv('PYPP_DEBUG'))
 ROOT_PATH = '/'
 PACKS_PATH = '/packs'
 SIMPLE_PATH = '/simple'
+
+DATA_PACKAGE = f'{__package__}.data'