+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 History
 -------
 
+ - Use form-based authentication for admin login and enable logout
+
 **2020-02-11 (0.2.0)**
  - Can now run without external package repository